Remove FacilityOrderStatusHelper
